try adding an alert task to influxdb
it doesn't work, because TasksApi.CreateTaskAsync always
fails with a timeout. the task is actually created, but the
game still crashes. this must have something to do, with the
game runtime's http settings, because the same code and flux
script work in a standalone .net 5 console app.

this was further complicated, because the game ships with an
old version of Newtonsoft.Json, which led to a different crash
whenever there is an http error, because the influxdb client uses
a method with a different signature in their newer version of
Newtonsoft.Json. i worked around it by changing that method and
compiling a custom version of InfluxDb.Client.Core.
the changed method was
InfluxDB.Client.Core.Exceptions.HttpException.Create
in InfluxException.cs. it can't use JObject.ContainsKey:
-                    if (errorBody.ContainsKey("message"))
+                    // if (errorBody.ContainsKey("message"))
+                    // {
+                    //     errorMessage = errorBody.GetValue("message").ToString();
+                    // }
+                    try
                     {
                         errorMessage = errorBody.GetValue("message").ToString();
                     }
+                    catch
+                    {
+                        errorMessage = null;
+                    }